frontend: split out the marine vectors presentation functionality
